Divide 'doc' dir into 'epydoc' and 'sphinx' to facilitate tries with both doc systems.
To that end, also add files needed to .gitignore and change epydoc build routines
(via external tools in PyDev) to find all neccessary config options in a config file
which is kept in the repo under the epydoc dir. 

Add a nice new CSS template for epydoc to make it more "Sphinx-like" as to
close the visual gap between the two systems (because, honestly, at the moment 
I cannot attest what benefits Sphinx has over Epydoc; besides the better initial 
impressions in the looks department anyway).